+++ kernel/rintegers.cc
@@ -85,7 +85,7 @@ void nrzPower (number a, int i, number * result)
 /*
  * create a number from int
  */
-number nrzInit (int i, const ring r)
+number nrzInit (long i, const ring r)
 {
   int_number erg = (int_number) omAllocBin(gmp_nrz_bin);
   mpz_init_set_si(erg, i);
